[BE] fix: 회원 ID 자료형 int로 수정 (S12P21A407-114)

+++ backend-spring/src/main/java/com/ssafy/em/auth/domain/entity/OAuth2CustomUser.java
@@ -14,7 +14,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class OAuth2CustomUser implements OAuth2User, Serializable {
 
-    private final Long id;
+    private final int id;
     private final String registrationId;
     private final Map<String, Object> attributes;
     private final List<GrantedAuthority> authorities;
